requirements.txt
python-docx
pyttsx3
